+++ packages/iconic/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added support for [`phosphor` icons](https://phosphoricons.com/), which are now bundled as part of the published content
+
 ## [1.2.1] - 2023-07-05
 
 ### Added